Refactor PEAR_UPLOAD_TMPDIR
This patch refactors the PEAR_UPLOAD_TMPDIR constant into a
configuration setting based on the app environment.
