+++ .gitignore
@@ -7,5 +7,8 @@
 # Static XML files for API access, generated by `bin/generate-rest.php`
 /public_html/rest/
 
+# Temporary generated application files (temporary uploaded files)
+/var/
+
 # Generated by Composer
 /vendor/
