+++ README.md
@@ -82,6 +82,7 @@ Source code of this application is structured in the following directories:
  ├─ src/             # Application source code classes
  ├─ templates/       # Application templates
  ├─ tests/           # Application automated tests
+ ├─ var/             # Temporary generated application files
  ├─ vendor/          # Dependencies generated by Composer
  ├─ phpunit.xml.dist # PHPUnit's default XML configuration
  ├─ .env.dist        # Distributed environment variables file for development
