+++ config/app_prod.php
@@ -27,4 +27,7 @@
 return [
     // REST static files directory
     'rest_dir' => isset($_SERVER['PECL_REST_DIR']) ? $_SERVER['PECL_REST_DIR'] : '/var/lib/peclweb/rest',
+
+    // Temporary directory for uploaded files
+    'tmp_uploads_dir' => isset($_SERVER['PECL_TMP_UPLOADS_DIR']) ? $_SERVER['PECL_TMP_UPLOADS_DIR'] : '/var/tmp/pear/uploads',
 ];
